new approach to req/resp/event types
each one is a bog standard module parameterised on
command/event and args/body poly params,
with a repeat for the optional cases

the code gen can then make
type t =
| CancelRequest of (Command.t, CancelArgs.t) RequestOpt
| NextRequest of (Command.t, NextArgs.t) Request
| ...etc

for all request types, similarly for all events and responses.

these can then be put into lists/hashtbls etc freely

NOTE we parameterise on command/event and hardcode the
ProtocolMessage_type because we dont want a dep of
dap_t on the auto-gen file, rather the auto-gen file
will be dependent on dap_t.