fixed the events/commands wrongness when read in from json
looks like the old enc for events/commands enums
couldnt actually create a specific enough type when used with
Json.destruct - the enum type got set to 'weak and then when combined
with the gadt ctor e.g. CancelRequest the weak resolved to whatever you
set it to.  What we want is that an error is raised somehow if we try
and destruct a cancel request that has the wrong command string.